+++ app/src/main/java/eu/kanade/tachiyomi/util/system/NotificationExtensions.kt
@@ -27,12 +27,7 @@ fun Context.notify(
 }
 
 fun Context.notify(id: Int, notification: Notification) {
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU &&
-        PermissionChecker.checkSelfPermission(
-            this,
-            Manifest.permission.POST_NOTIFICATIONS,
-        ) != PermissionChecker.PERMISSION_GRANTED
-    ) {
+    if (!canPostNotifications()) {
         return
     }
 
@@ -40,16 +35,19 @@ fun Context.notify(id: Int, notification: Notification) {
 }
 
 fun Context.notify(notificationWithIdAndTags: List<NotificationWithIdAndTag>) {
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU &&
-        PermissionChecker.checkSelfPermission(
-            this,
-            Manifest.permission.POST_NOTIFICATIONS,
-        ) != PermissionChecker.PERMISSION_GRANTED
-    ) {
+    if (!canPostNotifications()) {
         return
     }
 
     NotificationManagerCompat.from(this).notify(notificationWithIdAndTags)
+}
+
+fun Context.canPostNotifications(): Boolean {
+    return Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU ||
+        PermissionChecker.checkSelfPermission(
+            this,
+            Manifest.permission.POST_NOTIFICATIONS,
+        ) == PermissionChecker.PERMISSION_GRANTED
 }
 
 fun Context.cancelNotification(id: Int) {